search algorithm task 1

+++ 0x1E-search_algorithms/0-linear.c
@@ -23,7 +23,7 @@ int linear_search(int *array, size_t size, int value)
 	{
 		printf("value chacked array[%lu] = [%d]\n", i, array[i]);
 		if (array[i] == value)
-			return ((int)i);
+			return (i);
 	}
 	return (-1);
 }
